% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{as_mcmc_list}
\alias{as_mcmc_list}
\title{Convert an object to \code{mcmc.list}}
\usage{
as_mcmc_list(x)
}
\arguments{
\item{x}{An object of class \code{mcmc.list}, \code{mcmc}, \code{data.frame}, or \code{matrix}.}
}
\value{
An object of class \code{mcmc.list}.
}
\description{
Converts an \code{mcmc.list} / \code{mcmc} / \code{data.frame} / \code{matrix} into a \code{coda}-compatible \code{mcmc.list}.
}
\keyword{internal}
