% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_monitors_helpers.R
\name{as_mcmc_list_sop}
\alias{as_mcmc_list_sop}
\title{Convert various sample formats to coda::mcmc.list (optionally merging samples2)}
\usage{
as_mcmc_list_sop(samples, samples2 = NULL, drop_loglik = FALSE, thin = 1L)
}
\arguments{
\item{samples}{list/matrix/mcmc/list}

\item{samples2}{optional list/matrix/mcmc.list}

\item{drop_loglik}{drop log-likelihood columns}

\item{thin}{thinning for returned mcmc objects}
}
\value{
\code{coda::mcmc.list}
}
\description{
Convert various sample formats to coda::mcmc.list (optionally merging samples2)
}
\keyword{internal}
