% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runners.R
\name{build_conf_with_monitors}
\alias{build_conf_with_monitors}
\title{Build a fresh MCMC configuration with automatic monitors if missing}
\usage{
build_conf_with_monitors(model, monitors = NULL, opts = samOptiPro_options())
}
\arguments{
\item{model}{nimbleModel (ou compile; on accepte les deux)}

\item{monitors}{optional character vector; if NULL, auto-discover}

\item{opts}{samOptiPro_options()}
}
\description{
Build a fresh MCMC configuration with automatic monitors if missing
}
\keyword{internal}
