% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic_help.R
\name{compute_WAIC}
\alias{compute_WAIC}
\title{Compute WAIC from coda samples that include logLik[] columns}
\usage{
compute_WAIC(samples)
}
\arguments{
\item{samples}{coda::mcmc.list}
}
\description{
Compute WAIC from coda samples that include logLik[] columns
}
