% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_monitors_helpers.R
\name{.barrier_before_block}
\alias{.barrier_before_block}
\title{Temporarily switch nimble buildDir (clear compiled & GC)}
\usage{
.barrier_before_block(tag = "scalar_to_block")
}
\arguments{
\item{tag}{Character tag used in the temp dir name.}
}
\value{
TRUE (invisibly).
}
\description{
Temporarily switch nimble buildDir (clear compiled & GC)
}
\keyword{internal}
