% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitors.R
\name{.default_sanitize_roots}
\alias{.default_sanitize_roots}
\title{Discover default monitors from a nimble model (variable-level)}
\usage{
.default_sanitize_roots(x)
}
\arguments{
\item{model}{nimbleModel}

\item{opts}{list from samOptiPro_options(); champs supportés :
\itemize{
\item include_data      : logical, inclure les data nodes dans l'exploration (rarement utile)
\item include_logLik    : logical, inclure "logLik" dans la sélection par défaut
\item extra_monitors    : character(), variables supplémentaires à monitorer (samples)
\item extra_monitors2   : character(), variables supplémentaires à monitorer en monitors2 (samples2)
}}
}
\value{
character() de noms de variables sans indices (ex: "sd_proc","z","p","logLik")
}
\description{
Retourne des NOMS DE VARIABLES (sans indices) pour laisser le runner
poser ensuite les monitors correctement via conf$setMonitors / setMonitors2.
}
