% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_monitors_helpers.R
\name{.print_monitors}
\alias{.print_monitors}
\title{Pretty-print selected monitors (optional)}
\usage{
.print_monitors(title, nodes, thin = 1L)
}
\description{
Pretty-print selected monitors (optional)
}
\keyword{internal}
