% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_monitors_helpers.R
\name{.restore_builddir}
\alias{.restore_builddir}
\title{Restore previous nimble buildDir if it was changed}
\usage{
.restore_builddir()
}
\description{
Restore previous nimble buildDir if it was changed
}
\keyword{internal}
