% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_monitors_helpers.R
\name{.sop_expand_roots_to_nodes}
\alias{.sop_expand_roots_to_nodes}
\title{Expand many roots to nodes}
\usage{
.sop_expand_roots_to_nodes(mdl, roots)
}
\description{
Expand many roots to nodes
}
\keyword{internal}
