% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_monitors_helpers.R
\name{.sop_get_uncompiled_model}
\alias{.sop_get_uncompiled_model}
\title{Return the R-level (uncompiled) nimble model when given either R or compiled model}
\usage{
.sop_get_uncompiled_model(x)
}
\arguments{
\item{x}{A nimble model or a compiled model-like object.}
}
\value{
The R-level model environment if found, otherwise \code{x} unchanged.
}
\description{
Return the R-level (uncompiled) nimble model when given either R or compiled model
}
\keyword{internal}
