% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_monitors_helpers.R
\name{.sop_sanitize_roots}
\alias{.sop_sanitize_roots}
\title{Sanitize monitor roots (drop numerics & empties)}
\usage{
.sop_sanitize_roots(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector.
}
\description{
Sanitize monitor roots (drop numerics & empties)
}
\keyword{internal}
