% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{.sop_supports_derivs}
\alias{.sop_supports_derivs}
\title{Test différentiabilité/HMC avec ciblage par FAMILLES de nœuds}
\usage{
.sop_supports_derivs(model)
}
\description{
Test différentiabilité/HMC avec ciblage par FAMILLES de nœuds
}
