% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_monitors_helpers.R
\name{.us_expand_compact_node}
\alias{.us_expand_compact_node}
\title{Expand a compact node Character scalar, e.g. \code{p[1, 1:3]} into explicit indices (numeric only)}
\usage{
.us_expand_compact_node(node)
}
\arguments{
\item{node}{Single node string.}
}
\value{
Character vector of expanded nodes or \code{node} if not expandable.
}
\description{
Expand a compact node Character scalar, e.g. \code{p[1, 1:3]} into explicit indices (numeric only)
}
\keyword{internal}
