% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{merge_mcmc_samples}
\alias{merge_mcmc_samples}
\title{Merge \code{res$samples} and \code{res$samples2} into a single \code{mcmc.list}}
\usage{
merge_mcmc_samples(res)
}
\description{
Merge \code{res$samples} and \code{res$samples2} into a single \code{mcmc.list}
}
\keyword{internal}
