% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sampler.R
\name{monitors_sampled_from_conf}
\alias{monitors_sampled_from_conf}
\title{Extract sampled-only monitors from a configureMCMC object}
\usage{
monitors_sampled_from_conf(conf, monitors, strict = FALSE, keep_order = TRUE)
}
\arguments{
\item{conf}{nimble::MCMCconf returned by nimble::configureMCMC()}

\item{monitors}{character vector of base node names (your monitor vector)}

\item{strict}{logical; if TRUE, error if none are sampled}

\item{keep_order}{logical; if TRUE keep original order}
}
\value{
character vector monitors_sampled with attributes for diagnostics
}
\description{
Extract sampled-only monitors from a configureMCMC object
}
