% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sampler.R
\name{monitors_sampled_safe}
\alias{monitors_sampled_safe}
\title{One-pass build (uncompiled) + configureMCMC + monitors_sampled}
\usage{
monitors_sampled_safe(
  monitors,
  chain_id = 1L,
  strict = FALSE,
  keep_order = TRUE,
  buildDerivs = TRUE
)
}
\arguments{
\item{monitors}{character vector of requested monitors}

\item{chain_id}{integer(1)}

\item{strict}{logical}

\item{keep_order}{logical}

\item{buildDerivs}{logical passed to nimbleModel}
}
\value{
list(model=m, conf=conf.mcmc, monitors_sampled=..., chain_id=...)
}
\description{
One-pass build (uncompiled) + configureMCMC + monitors_sampled
}
