% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_generic.R
\name{normalize_constants_generic}
\alias{normalize_constants_generic}
\title{Normalize constants in a model-agnostic way}
\usage{
normalize_constants_generic(
  Const,
  n_rule = "min",
  pad_rule = "repeat_last",
  trim_rule = TRUE
)
}
\arguments{
\item{Const}{list of constants}

\item{n_rule}{"min"|"nyear"|"max"}

\item{pad_rule}{"repeat_last"|"zero"|"NA"}

\item{trim_rule}{logical (garder TRUE)}
}
\value{
constants list normalized (with integer 'n')
}
\description{
Normalize constants in a model-agnostic way
}
