% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runners.R
\name{run_baseline_coda}
\alias{run_baseline_coda}
\title{Run baseline et retourne un mcmc.list fusionne (samples + samples2)}
\usage{
run_baseline_coda(
  build_fn,
  niter,
  nburnin = floor(0.25 * niter),
  thin = 1L,
  monitors = NULL,
  nchains = 1L,
  drop_loglik = FALSE,
  opts = samOptiPro_options()
)
}
\description{
Run baseline et retourne un mcmc.list fusionne (samples + samples2)
}
\keyword{internal}
