% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runners.R
\name{run_baseline_config}
\alias{run_baseline_config}
\title{Run baseline RW/Slice with robust compile/run}
\usage{
run_baseline_config(
  build_fn,
  niter,
  nburnin = floor(0.25 * niter),
  thin = 1L,
  monitors = NULL,
  nchains = 1L,
  opts = samOptiPro_options()
)
}
\description{
Run baseline RW/Slice with robust compile/run
}
