% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runners.R
\name{run_hmc_all_nodes}
\alias{run_hmc_all_nodes}
\title{Run HMC/NUTS where possible; fallback samplers for uncovered nodes}
\usage{
run_hmc_all_nodes(
  build_fn,
  niter,
  nburnin = floor(0.25 * niter),
  thin = 1L,
  monitors = NULL,
  nchains = 1L,
  opts = samOptiPro_options()
)
}
\description{
Run HMC/NUTS where possible; fallback samplers for uncovered nodes
}
