% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sampler.R
\name{sampler_df}
\alias{sampler_df}
\title{Return a data.frame (name, type, target, scale) from a sampler container.
Accepts: samplerConf list (conf$getSamplers()) or other sampler-like lists.}
\usage{
sampler_df(samplers)
}
\description{
Return a data.frame (name, type, target, scale) from a sampler container.
Accepts: samplerConf list (conf$getSamplers()) or other sampler-like lists.
}
