% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sampler.R
\name{sampler_env_dump}
\alias{sampler_env_dump}
\title{Long data.frame of sampler env numeric fields}
\usage{
sampler_env_dump(cm)
}
\arguments{
\item{cm}{compiled MCMC}
}
\value{
data.frame(sampler_index, field, value)
}
\description{
Long data.frame of sampler env numeric fields
}
\keyword{internal}
