% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sampler.R
\name{sampler_functions_from_conf}
\alias{sampler_functions_from_conf}
\title{Build and compile an MCMC from a conf to access samplerFunctions.}
\usage{
sampler_functions_from_conf(conf, cmodel)
}
\description{
Build and compile an MCMC from a conf to access samplerFunctions.
}
