% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sampler.R
\name{sampler_scale}
\alias{sampler_scale}
\title{Extract a numeric scale proxy from a sampler object.
For RW: control$scale ; For slice: control$width (returned as 'scale').
Fallbacks: $scale (numeric or function) or environment(s)$scale.}
\usage{
sampler_scale(s)
}
\description{
Extract a numeric scale proxy from a sampler object.
For RW: control$scale ; For slice: control$width (returned as 'scale').
Fallbacks: $scale (numeric or function) or environment(s)$scale.
}
