% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sampler.R
\name{sampler_targets}
\alias{sampler_targets}
\title{Targets (node names) for each sampler in a MCMC configuration}
\usage{
sampler_targets(conf)
}
\arguments{
\item{conf}{a nimble MCMC configuration (from configureMCMC)}
}
\value{
character vector (length = length(conf$getSamplers()))
}
\description{
Targets (node names) for each sampler in a MCMC configuration
}
