% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utils.R
\name{time_it}
\alias{time_it}
\title{Internal helper to time an expression}
\usage{
time_it(expr)
}
\arguments{
\item{expr}{Expression to be evaluated.}
}
\value{
A list with two components:
\describe{
\item{result}{The evaluated value of \code{expr}.}
\item{runtime_s}{Elapsed time in seconds (numeric scalar).}
}
}
\description{
Evaluate an expression in the parent frame and return both the result
and the elapsed time in seconds. Intended for internal profiling of
MCMC runs and diagnostic steps.
}
\keyword{internal}
