% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitors.R
\name{.default_sanitize_roots}
\alias{.default_sanitize_roots}
\title{Internal helper to sanitize monitor root names}
\usage{
.default_sanitize_roots(x)
}
\arguments{
\item{x}{Character vector of candidate monitor roots.}
}
\value{
A character vector of cleaned, unique roots.
}
\description{
Internal helper to sanitize monitor root names
}
\keyword{internal}
