% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_monitors_helpers.R
\name{.sop_supports_derivs}
\alias{.sop_supports_derivs}
\title{Check if derivatives/HMC are supported for a nimble model}
\usage{
.sop_supports_derivs(model)
}
\arguments{
\item{model}{Nimble model.}
}
\value{
Logical.
}
\description{
Check if derivatives/HMC are supported for a nimble model
}
\keyword{internal}
