% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{plot_mcmc_histograms}
\alias{plot_mcmc_histograms}
\title{Histograms for ESS / CE / AE / Rhat (ggplot2)}
\usage{
plot_mcmc_histograms(
  samples,
  runtime_s,
  rhat_thresh = 1.01,
  bins = 30,
  log_x = TRUE
)
}
\description{
Histograms for ESS / CE / AE / Rhat (ggplot2)
}
